+++ app/src/main/res/anim/animation.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <scale
+        android:duration="300"
+        android:fromXScale="1.0"
+        android:fromYScale="1.0"
+        android:interpolator="@android:anim/decelerate_interpolator"
+        android:repeatCount="6"
+        android:repeatMode="reverse"
+        android:toXScale="1.25"
+        android:toYScale="1.25" />
+
+    <rotate
+        android:duration="300"
+        android:fromDegrees="0"
+        android:pivotX="50%"
+        android:pivotY="50%"
+        android:repeatCount="3"
+        android:repeatMode="reverse"
+        android:startOffset ="300"
+        android:toDegrees="360"
+        android:interpolator="@android:anim/decelerate_interpolator"/>
+</set>
